Moved the CLI parsing code into the `uniffi` crate.
This removes the clap dependency on `uniffi_bindgen` which is nice for
consumers that want to just use the Rust API.

This completes the migration of the CLI from `uniffi_bindgen` to
`uniffi`
