+++ uniffi_bindgen/Cargo.toml
@@ -10,17 +10,11 @@ license = "MPL-2.0"
 edition = "2021"
 keywords = ["ffi", "bindgen"]
 
-[[bin]]
-name = "uniffi-bindgen"
-path = "src/main.rs"
-required-features = ["clap"]
-
 [dependencies]
 anyhow = "1"
 askama = { version = "0.11", default-features = false, features = ["config"] }
 bincode = "1.3"
 camino = "1.0.8"
-clap = { version = "3.1", features = ["cargo", "std", "derive"], optional = true }
 fs-err = "2.7.0"
 glob = "0.3"
 goblin = "0.6"
@@ -33,6 +27,3 @@ toml = "0.5"
 weedle2 = { version = "4.0.0", path = "../weedle2" }
 uniffi_meta = { path = "../uniffi_meta", version = "=0.22.0" }
 uniffi_testing = { path = "../uniffi_testing", version = "=0.22.0" }
-
-[features]
-default = ["clap"]
