--- uniffi_bindgen/src/main.rs
@@ -1,9 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-use anyhow::Result;
-
-fn main() -> Result<()> {
-    uniffi_bindgen::run_main()
-}
